requirements.txt
alabaster==0.7.12
attrs==21.4.0
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.0.12
coverage==6.4.1
distlib==0.3.4
docutils==0.18.1
filelock==3.7.1
furo==2022.6.21
gitdb==4.0.9
GitPython==3.1.27
identify==2.5.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
Jinja2==3.1.2
markdown-it-py==2.1.0
MarkupSafe==2.1.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
myst-parser==0.18.0
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
py==1.11.0
pycommando==1.2.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinx-autodoc-typehints==1.18.3
sphinx-basic-ng==0.0.1a11
sphinx-copybutton==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==2.0.1
tox==3.25.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
zipp==3.8.0
